+++ src/utils/auth/duplicate-validation.ts
@@ -3,7 +3,7 @@
 import { TABLES } from '@/constants/tables.constant';
 import { ERROR_MESSAGES } from '@/constants/auth.messages.constant';
 import { User } from '@/types/supabase.type';
-import { getSingleData } from '@/services/common.server.dto';
+import { getSingleData } from '@/apis/common-server.api';
 
 export const duplicateEmailValidation = async (email: string) => {
   const { data, error } = await getSingleData<User>({
